Add build flag to disable TFGen
I'd like to make a tool that makes it possible to build tests that compare function outputs from multiple Haskell backends against each other, in order to ensure consistent behavior on multiple backends. Currently I can't use Arbitrary for this tool, because I need to ensure that the randomized output is the same on all backends, Adding a build flag to disable TFGen should make this possible, albeit somewhat hacky since I have to somehow ensure that QuickCheck was always built with this flag. Perhaps it would be better to instead make it possible to choose exactly which generator to use? This could be done, for example, by using a SomeGen existential, but tha would require the use of language extensions. Is there another way to allow this?