Write specialized instance for sequences
Replace list-based `Arbitrary` instance with one using
primitives from `Data.Sequence`. This avoids the inefficiency
of `fromList` in general, and takes advantage of fast splitting
and appending.

Note: there may well be a way to improve shrinking further;
this commit simply translates the version for lists into
the language of sequences.