introduce 'assertWith' for better debugging on monadic assertions
The current 'assert' is a bit rough in the sense that it doesn't give any feedback other than something failed. In particular, if a test makes several assertions it can be tricky to figure out which one failed and for which reason. 'assertWith' makes it possible to specify an error message that is shown only on failure (piggy back on 'counterexample') but, clearly shows the result of the assertion sofor an easier debugging output.