Made instances only exist when Newline type exists
Newline and NewlineMode were added in commits
7b067f2d08d4968ab323405828fa2f053964ffb0
and
6bc00b292d2de418b87de44ffb9ad00541b80b43
to the GHC repository.

Searching through Hackage shows that this corresonds to base 4.4.0.0.
In Base 4.4.0.0, these types are only defined when __NHC__ and
__HUGS__ are not defined.
Thus, the constraint
MIN_VERSION_base(4,4,0) && !defined(__NHC__) && !defined(__HUGS__)

As of Base 4.7.0.0, these NHC and HUGS constraints were dropped, and to
maintain maximum compatiblity with base (Arbitrary Newline is defined
if and only if Newline is defined)
The alternative contraint
MIN_VERSION_base(4,7,0) was provided.

This should fix the non-GHC compatibility problems.