Get rid of magic behaviour where properties without quantification were only tested once (e.g.: quickCheck True).
This behaviour has led to strange behaviour in the past and proved rather difficult to get right. Here is the latest case:

(1) quickCheck . checkCoverage $ coverTable ”A” [(”a”,1)] True
(2) quickCheck . checkCoverage $ coverTable ”A” [(”a”,1)] (\() -> True)

Property (1) succeeds, but property (2) continues testing forever.

It seems simpler to just remove this, and people can use `once` or `withMaxSuccess 1` to limit the number of tests explicitly. In practice, the only properties affected will be those of type IO Bool, which will now be tested 100 times rather than once.